Refactor game logic into components
---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/Mst-Takayama/typing_game?shareId=XXXX-XXXX-XXXX-XXXX).
